+++ Software/wittypi/utilities.sh
@@ -148,7 +148,7 @@ if [ -z ${I2C_MC_ADDRESS+x} ]; then
 
   TIME_UNKNOWN=0
 
-  SOFTWARE_VERSION='5.51'
+  SOFTWARE_VERSION='5.52'
 
   readonly LOCAL_TZ='Europe/London'
 fi